# api/openapi.yaml
openapi: 3.0.3
info:
  title: Firebase Authentication Go Proxy API
  description: |-
    This is the Firebase Authentication Go Proxy API documentation.
  termsOfService: https://example.com
  contact:
    email: sample@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8080/api
tags:
  - name: auth
    description: 認証
paths:
  /v1/auth/signup:
    post:
      tags:
        - auth
      summary: サインアップ
      description: ユーザーを登録する
      operationId: v1AuthSignUp
      security:
        - apiKey: []      
      requestBody:
        required: true
        description: サインアップリクエストボディ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1AuthSignUpRequestSchema'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /v1/auth/signin:
    post:
      tags:
        - auth
      summary: サインイン
      description: ユーザーを認証する
      operationId: v1AuthSignIn
      requestBody:
        required: true
        description: サインインリクエストボディ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1AuthSignInRequestSchema'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1AuthSignInResponseSchema'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /v1/auth/verify:
    get:
      tags:
        - auth
      summary: 検証
      description: 検証を行う
      operationId: v1AuthVerify
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /v1/auth/refresh:
    post:
      tags:
        - auth
      summary: リフレッシュ
      description: リフレッシュトークンによりアクセストークンを更新する
      operationId: v1AuthRefresh
      requestBody:
        required: true
        description: リフレッシュリクエストボディ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1AuthRefreshRequestSchema'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1AuthRefreshResponseSchema'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /v1/auth/password:
    put:
      tags:
        - auth
      summary: パスワード更新
      description: パスワードを更新する
      operationId: v1AuthChangePassword
      security:
        - Bearer: []
      requestBody:
        required: true
        description: パスワード更新リクエストボディ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1AuthChangePasswordRequestSchema'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /v1/auth/signout:
    get:
      tags:
        - auth
      summary: サインアウト
      description: サインアウトする
      operationId: v1AuthSignOut
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /v1/auth:
    delete:
      tags:
        - auth
      summary: リサイン(退会)
      description: リサイン(退会)する
      operationId: v1AuthResign
      security:
        - Bearer: []
      requestBody:
        required: true
        description: サインインリクエストボディ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1AuthResignRequestSchema'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token
  schemas:
    V1AuthSignUpRequestSchema:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          format: email
          example: sample@example.com
        password:
          type: string
          description: パスワード
          example: password
      required:
        - email
        - password
    V1AuthSignInRequestSchema:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          format: email
          example: sample@example.com
        password:
          type: string
          description: パスワード
          example: password
      required:
        - email
        - password
    V1AuthSignInResponseSchema:
      type: object
      properties:
        idToken:
          type: string
          description: IDトークン
        refreshToken:
          type: string
          description: リフレッシュトークン
      required:
        - idToken
        - refreshToken
    V1AuthRefreshRequestSchema:
      type: object
      properties:
        refreshToken:
          type: string
          description: リフレッシュトークン
      required:
        - refreshToken
    V1AuthRefreshResponseSchema:
      type: object
      properties:
        idToken:
          type: string
          description: IDトークン
      required:
        - idToken
    V1AuthChangePasswordRequestSchema:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          format: email
          example: sample@example.com
        oldPassword:
          type: string
          description: 旧パスワード
          example: password
        newPassword:
          type: string
          description: 新パスワード
          example: new_password
      required:
        - email
        - oldPassword
        - newPassword
    V1AuthResignRequestSchema:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          format: email
          example: sample@example.com
        password:
          type: string
          description: パスワード
          example: password
      required:
        - email
        - password