# internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package api

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	BearerScopes = "Bearer.Scopes"
	ApiKeyScopes = "apiKey.Scopes"
)

// V1AuthChangePasswordRequestSchema defines model for V1AuthChangePasswordRequestSchema.
type V1AuthChangePasswordRequestSchema struct {
	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// NewPassword 新パスワード
	NewPassword string `json:"newPassword"`

	// OldPassword 旧パスワード
	OldPassword string `json:"oldPassword"`
}

// V1AuthRefreshRequestSchema defines model for V1AuthRefreshRequestSchema.
type V1AuthRefreshRequestSchema struct {
	// RefreshToken リフレッシュトークン
	RefreshToken string `json:"refreshToken"`
}

// V1AuthRefreshResponseSchema defines model for V1AuthRefreshResponseSchema.
type V1AuthRefreshResponseSchema struct {
	// IdToken IDトークン
	IdToken string `json:"idToken"`
}

// V1AuthResignRequestSchema defines model for V1AuthResignRequestSchema.
type V1AuthResignRequestSchema struct {
	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// Password パスワード
	Password string `json:"password"`
}

// V1AuthSignInRequestSchema defines model for V1AuthSignInRequestSchema.
type V1AuthSignInRequestSchema struct {
	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// Password パスワード
	Password string `json:"password"`
}

// V1AuthSignInResponseSchema defines model for V1AuthSignInResponseSchema.
type V1AuthSignInResponseSchema struct {
	// IdToken IDトークン
	IdToken string `json:"idToken"`

	// RefreshToken リフレッシュトークン
	RefreshToken string `json:"refreshToken"`
}

// V1AuthSignUpRequestSchema defines model for V1AuthSignUpRequestSchema.
type V1AuthSignUpRequestSchema struct {
	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// Password パスワード
	Password string `json:"password"`
}

// V1AuthResignJSONRequestBody defines body for V1AuthResign for application/json ContentType.
type V1AuthResignJSONRequestBody = V1AuthResignRequestSchema

// V1AuthChangePasswordJSONRequestBody defines body for V1AuthChangePassword for application/json ContentType.
type V1AuthChangePasswordJSONRequestBody = V1AuthChangePasswordRequestSchema

// V1AuthRefreshJSONRequestBody defines body for V1AuthRefresh for application/json ContentType.
type V1AuthRefreshJSONRequestBody = V1AuthRefreshRequestSchema

// V1AuthSignInJSONRequestBody defines body for V1AuthSignIn for application/json ContentType.
type V1AuthSignInJSONRequestBody = V1AuthSignInRequestSchema

// V1AuthSignUpJSONRequestBody defines body for V1AuthSignUp for application/json ContentType.
type V1AuthSignUpJSONRequestBody = V1AuthSignUpRequestSchema